Convert ipynb to org to handle recursion
Note that we had to extract the basename, since sed is unhappy with
replacement strings that contain /s, since it treats everything after
them as flags. When doing the recursive version of this, many of the
files will have paths, and we handle this by using the basename function.

We also add a force flag, though we don't enable calling it in emacs.
Instead I'll run

find . -name "*.org", with -delete when I'm ready to remove the files.
This feels safer to me than having something I could accidentally run.